+++ includes/core.php
@@ -31,7 +31,7 @@ function setup() {
 function i18n() {
 	$locale = apply_filters( 'plugin_locale', get_locale(), 'cloudflare-access-sso' );
 	load_textdomain( 'cloudflare-access-sso', WP_LANG_DIR . '/cloudflare-access-sso/cloudflare-access-sso-' . $locale . '.mo' );
-	load_plugin_textdomain( 'cloudflare-access-sso', false, plugin_basename( CLOUDFLARE_ACCESS_SSO_PLUGIN_PATH ) . '/languages/' );
+	load_plugin_textdomain( 'cloudflare-access-sso', false, plugin_basename( CF_ACCESS_SSO_PLUGIN_PATH ) . '/languages/' );
 }
 
 /**
@@ -40,6 +40,8 @@ function i18n() {
  * @return void
  */
 function init() {
+
+	do_action( 'cloudflare_access_sso_plugin_pre_init' );
 
 	if ( class_exists( '\CloudflareAccessSSO\Plugin' ) ) {
 		\CloudflareAccessSSO\Plugin::get_instance()->setup();
@@ -64,6 +66,6 @@ function activate() {
  * @return void
  */
 function deactivate() {
-	// Flush cache to remove Cloudflare Certificates
+	// Flushing the cache removes stored Cloudflare Certificates
 	wp_cache_flush();
 }